+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 FEATURES
 
 * **New Resource:** `metabase_permissions_group`
+* **New Data Source:** `metabase_permissions_group`
 
 ## 0.3.0 (October 11, 2022)